+++ DataAccess/Repository/CategoryRepository.cs
@@ -9,12 +9,12 @@ namespace DataAccess.Repository
         public async Task Add(Category category)
             => await CategoryDAO.Add(category);
 
-        public async Task<bool> CheckProductExistsByCateId(int categoryId)
+        public async Task<bool> CheckProductExistsByCateId(long categoryId)
             => await CategoryDAO.CheckProductExistsByCateId(categoryId);
 
         public async Task<List<Category>> GetAll() => await CategoryDAO.GetAll();
 
-        public async Task<Category?> FindById(int categoryId)
+        public async Task<Category?> FindById(long categoryId)
             => await CategoryDAO.FindById(categoryId);
 
         public async Task<Category?> FindByName(string name)
