+++ DataAccess/Repository/IRepository/ICouponRepository.cs
@@ -5,7 +5,7 @@ namespace DataAccess.Repository.IRepository
     public interface ICouponRepository
     {
         Task<List<Coupon>> GetAll();
-        Task<Coupon?> FindById(int id);
+        Task<Coupon?> FindById(long id);
         Task<Coupon?> FindByCode(string code);
         Task Add(Coupon coupon);
         Task Update(Coupon coupon);
