+++ DataAccess/Repository/IRepository/IReportProductReviewRepository.cs
@@ -6,8 +6,8 @@ namespace DataAccess.Repository.IRepository
     public interface IReportProductReviewRepository
     {
         Task<List<ReportProductReview>> GetAll();
-        Task<ReportProductReview?> FindById(int id);
-        Task<ReportProductReview?> FindByUserIdAndProductReviewId(int userId, int productReviewId);
+        Task<ReportProductReview?> FindById(long id);
+        Task<ReportProductReview?> FindByUserIdAndProductReviewId(long userId, long productReviewId);
         Task Add(ReportProductReview reportProductReview);
         Task Update(ReportProductReview reportProductReview);
     }
