+++ DataAccess/Repository/UserRepository.cs
@@ -12,7 +12,7 @@ namespace DataAccess.Repository
     public class UserRepository : IUserRepository
     {
         public async Task<List<User>> GetAll() => await UserDAO.GetAll();
-        public async Task<User?> FindById(int id) => await UserDAO.FindById(id);
+        public async Task<User?> FindById(long id) => await UserDAO.FindById(id);
         public async Task<User?> FindByEmail(string email) => await UserDAO.FindByEmail(email);
         public async Task Add(User user) => await UserDAO.Add(user);
         public async Task Update(User user) =>await UserDAO.Update(user);
